// src/backend/chat/twitch-slash-commands.d.ts
export type TwitchSlashCommandValidationResult<Args extends unknown[] = unknown[]> = {
    success: true;
    args: Args;
} | {
    success: false;
    foundCommand?: boolean;
    errorMessage: string;
};

export type TwitchSlashCommandHandler<Args extends unknown[] = unknown[]> = {
    commands: string[];
    validateArgs(rawArgs: string[]): TwitchSlashCommandValidationResult<Args>;
    handle(args: Args, sendAsBot?: boolean): boolean | PromiseLike<boolean>
};